OAK-11721: compatibility mode test
